Rename: 예외 클래스를 global 폴더로 이동
