+++ src/main/java/com/jm0514/myboard/auth/AuthenticationPrincipalArgumentResolver.java
@@ -1,6 +1,6 @@
 package com.jm0514.myboard.auth;
 
-import com.jm0514.myboard.advice.AuthException;
+import com.jm0514.myboard.global.exception.AuthException;
 import com.jm0514.myboard.auth.domain.JwtProvider;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
