+++ src/main/java/com/jm0514/myboard/global/exception/AuthException.java
@@ -1,6 +1,4 @@
-package com.jm0514.myboard.advice;
-
-import com.jm0514.myboard.global.exception.ExceptionStatus;
+package com.jm0514.myboard.global.exception;
 
 public class AuthException extends RuntimeException{
