+++ src/main/java/com/jm0514/myboard/member/exception/NotFoundMemberException.java
@@ -1,6 +1,6 @@
 package com.jm0514.myboard.member.exception;
 
-import com.jm0514.myboard.advice.BadRequestException;
+import com.jm0514.myboard.global.exception.BadRequestException;
 import com.jm0514.myboard.global.exception.ExceptionStatus;
 
 public class NotFoundMemberException extends BadRequestException {
