+++ src/main/java/com/jm0514/myboard/member/service/MemberService.java
@@ -1,6 +1,6 @@
 package com.jm0514.myboard.member.service;
 
-import com.jm0514.myboard.advice.BadRequestException;
+import com.jm0514.myboard.global.exception.BadRequestException;
 import com.jm0514.myboard.auth.dto.AuthInfo;
 import com.jm0514.myboard.member.domain.Member;
 import com.jm0514.myboard.member.dto.MemberInfoRequestDto;
