+++ src/test/java/com/jm0514/myboard/auth/domain/JwtProviderTest.java
@@ -1,6 +1,6 @@
 package com.jm0514.myboard.auth.domain;
 
-import com.jm0514.myboard.advice.AuthException;
+import com.jm0514.myboard.global.exception.AuthException;
 import com.jm0514.myboard.auth.dto.AuthInfo;
 import io.jsonwebtoken.Header;
 import io.jsonwebtoken.Jwts;
